uhttp: Support abstract sockets
Abstract socket are not bound to filesystem pathname, and do not use
permissions or ownership. We like to use an abstract socket for
accessing images locally.

Skip unlinking of abstract socket (starting with "\0"), since there is
noting on the filesystem to remove, and os.unlink() does not like
pathname with embedded nulls.

Using empty socket address will bind the server to random abstract
socket picked by the kernel. This is useful for testing on a host with a
running daemon.

+++ daemon/ovirt_imageio_daemon/uhttp.py
@@ -62,14 +62,22 @@ class UnixWSGIServer(wsgi.WSGIServer):
 
     def server_bind(self):
         """
-        Override to remove existing socket.
+        Override to remove existing socket for pathname sockets, and support
+        random abstract sockets.  See unix(7) for details.
         """
-        try:
-            os.unlink(self.server_address)
-        except OSError as e:
-            if e.errno != errno.ENOENT:
-                raise
+        if self.server_address == "":
+            # User wants to bind to a random abstract socket.
+            self.socket.setsockopt(socket.SOL_SOCKET, socket.SO_PASSCRED, 1)
+        elif self.server_address[0] != "\0":
+            # A pathname socket must be removed before binding.
+            try:
+                os.unlink(self.server_address)
+            except OSError as e:
+                if e.errno != errno.ENOENT:
+                    raise
+
         self.socket.bind(self.server_address)
+        self.server_address = self.socket.getsockname()
         self.setup_environ()
 
     def get_request(self):
